+++ framework-crates/objc2-core-services/src/files.rs
@@ -7,7 +7,8 @@ use crate::UniChar;
 // act as-if CoreServices is the one to define this type.
 //
 // NOTE: This is marked __attribute__((aligned(2), packed)), but that's
-// unnecessary, since the layout of the .
+// unnecessary, since the alignment of u16/UniChar, and thereby the struct
+// itself, is already 2.
 #[repr(C)]
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct HFSUniStr255 {
@@ -26,4 +27,14 @@ unsafe impl Encode for HFSUniStr255 {
 #[cfg(feature = "objc2")]
 unsafe impl RefEncode for HFSUniStr255 {
     const ENCODING_REF: Encoding = Encoding::Pointer(&Self::ENCODING);
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn alignment() {
+        assert_eq!(core::mem::align_of::<HFSUniStr255>(), 2);
+    }
 }