+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -27,6 +27,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   takes a nullable file name.
 * **BREAKING**: The media selection option on `AVAssetVariantQualifier` is now nullable.
 
+### Fixed
+* **BREAKING**: Fix structs with packed alignment by marking them `#[repr(packed(...))]`.
+
 ## [0.3.1] - 2025-04-19
 [0.3.1]: https://github.com/madsmtm/objc2/compare/frameworks-0.3.0...frameworks-0.3.1